// libs/api/produttori-service/src/lib/model/azienda-update-dto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AziendaUpdateDTO { 
    id?: string;
    nome: string;
    numeroPubblico: string;
    email: string;
    produttore: string;
    comune: string;
    provincia: string;
}